move plugin registration function
