+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/index.ts
@@ -5,11 +5,43 @@
  * 2.0.
  */
 
-export { getDocumentByIdTool } from './get_document_by_id';
-export { getIndexMappingsTool } from './get_index_mapping';
-export { listIndicesTool } from './list_indices';
-export { indexExplorerTool } from './index_explorer';
-export { generateEsqlTool } from './generate_esql';
-export { executeEsqlTool } from './execute_esql';
-export { searchTool } from './search';
-export { createVisualizationTool } from './create_visualization';
+import type { CoreSetup } from '@kbn/core-lifecycle-server';
+import type { BuiltinToolDefinition } from '@kbn/onechat-server';
+import type {
+  AgentBuilderPlatformPluginStart,
+  PluginSetupDependencies,
+  PluginStartDependencies,
+} from '../types';
+import { getDocumentByIdTool } from './get_document_by_id';
+import { getIndexMappingsTool } from './get_index_mapping';
+import { listIndicesTool } from './list_indices';
+import { indexExplorerTool } from './index_explorer';
+import { generateEsqlTool } from './generate_esql';
+import { executeEsqlTool } from './execute_esql';
+import { searchTool } from './search';
+import { createVisualizationTool } from './create_visualization';
+
+export const registerTools = ({
+  coreSetup,
+  setupDeps,
+}: {
+  coreSetup: CoreSetup<PluginStartDependencies, AgentBuilderPlatformPluginStart>;
+  setupDeps: PluginSetupDependencies;
+}) => {
+  const { onechat } = setupDeps;
+
+  const tools: Array<BuiltinToolDefinition<any>> = [
+    searchTool(),
+    getDocumentByIdTool(),
+    executeEsqlTool(),
+    generateEsqlTool(),
+    getIndexMappingsTool(),
+    listIndicesTool(),
+    indexExplorerTool(),
+    createVisualizationTool(),
+  ];
+
+  tools.forEach((tool) => {
+    onechat.tools.register(tool);
+  });
+};
